requirements.txt
absl-py==0.10.0
astor==0.8.1
# Editable install with no version control (baselines==0.1.6)
-e /home/lunar3/Desktop/PythonProjects/star/baselines
click==7.1.2
cloudpickle==1.2.2
# Editable Git install with no remote (customFiles==0.0.1)
-e /home/lunar3/Desktop/PythonProjects/star
future==0.18.2
gast==0.4.0
google-pasta==0.2.0
grpcio==1.32.0
gym==0.15.3
h5py==2.10.0
importlib-metadata==2.0.0
joblib==0.16.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.2.2
numpy==1.19.2
opencv-python==4.4.0.44
protobuf==3.13.0
pyglet==1.3.2
PyYAML==5.3.1
RLSTaR==0.0.1
RLSTaR.customFiles==0.0.1
scipy==1.5.2
six==1.15.0
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
tqdm==4.49.0
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.2.0
